+++ mvc/src/main/java/com/interface21/webmvc/servlet/DispatcherServlet.java
@@ -1,11 +1,7 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet;
 
 import com.interface21.webmvc.HandlerMappings;
 import com.interface21.webmvc.RequestMappingHandlerAdapter;
-import com.interface21.webmvc.servlet.HandlerAdapter;
-import com.interface21.webmvc.servlet.HandlerMapping;
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.View;
 import com.interface21.webmvc.servlet.mvc.HandlerAdapters;
 import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
 import com.interface21.webmvc.servlet.view.JsonViewResolver;
@@ -15,23 +11,32 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class DispatcherServlet extends HttpServlet {
 
     private static final long serialVersionUID = 1L;
-    private static final String BASE_PACKAGE = DispatcherServlet.class.getPackageName();
     private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
 
+    private final List<String> scanPackages;
     private HandlerMappings handlerMappings;
     private HandlerAdapters handlerAdapters;
     private ViewResolvers viewResolvers;
 
+    public DispatcherServlet(String... basePackage) {
+        scanPackages = new ArrayList<>();
+        scanPackages.add(getClass().getPackageName());
+        scanPackages.addAll(Arrays.stream(basePackage).toList());
+    }
+
     @Override
     public void init() {
         handlerMappings = new HandlerMappings();
-        handlerMappings.addHandlerMapping(new AnnotationHandlerMapping(BASE_PACKAGE));
+        handlerMappings.addHandlerMapping(new AnnotationHandlerMapping(scanPackages));
         handlerAdapters = new HandlerAdapters();
         handlerAdapters.addHandlerAdapter(new RequestMappingHandlerAdapter());
         viewResolvers = new ViewResolvers();
@@ -52,8 +57,7 @@ public class DispatcherServlet extends HttpServlet {
             HandlerAdapter adapter = handlerAdapters.getHandlerAdapter(handler);
             ModelAndView modelAndView = adapter.handle(handler, request, response);
 
-            String viewName = modelAndView.getViewName();
-            View view = viewResolvers.resolveViewName(viewName);
+            View view = viewResolvers.resolveViewName(modelAndView.getViewName());
             view.render(modelAndView.getModel(), request, response);
         } catch (Throwable e) {
             log.error("Exception : {}", e.getMessage(), e);
